Fix remote cwd test compile
