+++ Packages/macOS/CmuxControlSocket/Tests/CmuxControlSocketTests/FakeSurfaceControlCommandContext.swift
@@ -1,0 +1,17 @@
+import Foundation
+@testable import CmuxControlSocket
+
+@MainActor
+final class FakeSurfaceControlCommandContext: ControlCommandContext {
+    var reportPWDResolution: ControlSurfaceReportPWDResolution = .recorded(surfaceID: UUID())
+    var reportedPWD: (workspaceID: UUID, requestedSurfaceID: UUID?, path: String)?
+
+    func controlSurfaceReportPWD(
+        workspaceID: UUID,
+        requestedSurfaceID: UUID?,
+        path: String
+    ) -> ControlSurfaceReportPWDResolution {
+        reportedPWD = (workspaceID, requestedSurfaceID, path)
+        return reportPWDResolution
+    }
+}
